feat: moved the logic from the CardsSets to Set controller
