+++ src/main/java/com/example/rebrain/controllers/SetController.java
@@ -1,18 +1,14 @@
 package com.example.rebrain.controllers;
 
-import com.example.rebrain.dto.*;
+import com.example.rebrain.dto.CardDto;
+import com.example.rebrain.dto.CardsSetsGetDto;
 import com.example.rebrain.dto.SetDto;
-import com.example.rebrain.dto.SetDto;
-import com.example.rebrain.entity.SetEntity;
-import com.example.rebrain.entity.SetEntity;
+import com.example.rebrain.entity.CardEntity;
 import com.example.rebrain.entity.SetEntity;
 import com.example.rebrain.exception.ObjectNotFoundException;
-import com.example.rebrain.mapper.*;
-import com.example.rebrain.mapper.SetMapper;
-import com.example.rebrain.mapper.SetMapper;
+import com.example.rebrain.mapper.CardMapper;
 import com.example.rebrain.mapper.SetMapper;
 import lombok.AllArgsConstructor;
-import com.example.rebrain.entity.SetEntity;
 import com.example.rebrain.services.SetService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -31,15 +27,11 @@ public class SetController {
 
     @PostMapping
     public ResponseEntity<SetDto> createSet(@RequestBody SetDto setDto) {
-        try {
-            SetEntity setEntity = SetMapper.toEntity(setDto);
-            SetEntity createdSet = setService.create(setEntity);
-            SetDto createdToDto = SetMapper.toDto(createdSet);
-            URI location = URI.create("/sets/" + createdToDto.getId());
-            return ResponseEntity.created(location).body(createdToDto);
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body(null);
-        }
+        SetEntity setEntity = SetMapper.toEntity(setDto);
+        SetEntity createdSet = setService.create(setEntity);
+        SetDto createdToDto = SetMapper.toDto(createdSet);
+        URI location = URI.create("/sets/" + createdToDto.getId());
+        return ResponseEntity.created(location).body(createdToDto);
     }
 
     @GetMapping
@@ -65,6 +57,17 @@ public class SetController {
         } catch (Exception e) {
             return ResponseEntity.badRequest().body(null);
         }
+    }
+
+    @GetMapping("/{id}/cards")
+    public ResponseEntity<CardsSetsGetDto> getCardsBySetId(@PathVariable Integer id) {
+        List<CardEntity> cardEntities = setService.getCardsBySetId(id);
+        List<CardDto> cardDtos = cardEntities.stream()
+                .map(CardMapper::toDto)
+                .collect(Collectors.toList());
+        SetEntity setEntity = setService.getOne(id);
+        SetDto setDto = SetMapper.toDto(setEntity);
+        return ResponseEntity.ok(new CardsSetsGetDto(id, setDto.getTitle(), setDto.getDescription(), cardDtos));
     }
 
     @PatchMapping("/{id}")
