+++ src/main/java/com/example/rebrain/dto/CardPostDto.java
@@ -1,0 +1,13 @@
+package com.example.rebrain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class CardPostDto {
+    private Integer setId;
+    private CardDto card;
+}
